Make Source static initialization more portable
Using nudetect in a Python script in a different directory causes a
FileNotFoundException, since Pandas looks for the isotope and x-ray
source data in the current directory, instead of in the nudetect
installation directory.

This fixes the problem by providing Pandas with an absolute path to the
isotope and x-ray source data files, increasing nudetect's portability.